requirements.txt
torch==1.13.1
torchvision==0.14.1
pytorch-lightning==1.9.5
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
Pillow>=9.4.0
PyYAML>=6.0
tqdm>=4.64.0
mediapipe>=0.9.0
onnx>=1.13.0
onnxruntime>=1.13.0
tensorflow>=2.11.0
ptflops>=0.6.9
matplotlib>=3.6.0
seaborn>=0.12.0
scipy>=1.10.0
pytest>=7.2.0
pytest-cov>=4.0.0
opencv-python>=4.7.0
